pyproject.toml
[project]
name = "ngdai"
version = "0.1.0"
description = "Wissensbasierte Analyseplattform fuer die deutsche Netzentgeltregulierung"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",

    # CLI
    "typer[all]>=0.12.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",

    # LLM
    "anthropic>=0.40.0",

    # Embeddings (Phase 2, but install early)
    # "sentence-transformers>=3.0.0",

    # Fuzzy Matching
    "rapidfuzz>=3.9.0",

    # Configuration
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",

    # Redis
    "redis>=5.0.0",

    # Utilities
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]

[project.scripts]
ngdai = "ngdai.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ngdai"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
